Move push-notification hook handler out of cmux.swift
workflow-guard-tests failed the Swift file length budget: cmux.swift was
exactly at budget before this branch and the new handler added 104 lines.
Pure code motion into CLI/CMUXCLI+ClaudePushNotificationHook.swift (under
the 500-line tracking threshold); cmux.swift is back to its budgeted
34499 lines. Widened only the helpers the moved code calls from private
to internal.
